move components dir into the src
