+++ src/app/error.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useEffect } from 'react';
-import Container from './components/ui/container';
-import { FadeIn, FadeInStagger } from './components/fade-in';
-import Button from './components/ui/Button';
+import Container from '@/src/components/ui/container';
+import { FadeIn, FadeInStagger } from '@/src/components/fade-in';
+import Button from '@/src/components/ui/Button';
 
 type ErrorPageProps = {
   error: Error;
