+++ src/components/canvas-gradient-inner.tsx
