+++ src/components/icons/drawing-icon.tsx
