+++ src/components/icons/feature3.tsx
