+++ src/components/icons/hero-animation.tsx
