+++ src/components/icons/logo-secondary.tsx
