+++ src/components/icons/magnify-icon.tsx
