+++ src/components/icons/minimized/featur.tsx
