+++ src/components/icons/minimized/feature1.tsx
