+++ src/components/icons/minimized/feature2.tsx
