+++ src/components/icons/minimized/feature3.tsx
