+++ src/components/info-card.tsx
