+++ src/components/quick-view.tsx
